Use DynamoDB Local and Tomcat Maven Plugin for local development.
System properties are set in dynamodb-geo-server/pom.xml in the tomcat plugin configuration. If is_local is defined the servlet will try to connect to a DynamoDB local instance on the default port. You can also use this to run the tomcat instance locally instead of using elastic beanstalk.

The server can be run by executing 'mvn tomcat7:run' and will listen on localhost:9090